+++ pages/prof/progress.vue
@@ -1,0 +1,87 @@
+<template>
+  <dashboardsProf-dashboard nav="progress">
+    <div class="py-6">
+      <div class="max-w-8xl mx-auto px-4 sm:px-6 md:px-8">
+        <h1 class="text-2xl font-semibold text-gray-900">
+          Aktueller Stand &amp; Zukunftsaussichten
+        </h1>
+      </div>
+      <div class="max-w-8xl mx-auto px-4 sm:px-6 md:px-8 pt-4">
+        <!-- Replace with your content -->
+        <div class="py-4 grid grid-cols-2 grid-flow-row gap-4">
+          <div class="bg-white shadow sm:rounded-lg">
+            <div class="px-4 py-5 sm:p-6">
+              <h3 class="text-lg leading-6 font-medium text-gray-900">
+                Sebastian - Prof Dashboard, LA &amp; EDM
+              </h3>
+              <div class="mt-2 text-gray-600">
+                <p>
+                  Was wurde bereits umgesetzt:
+                  <ul class="list-disc list-inside">
+                    <li>Gerüst der WebApp mit unterschiedlichen Bereichen und einem einheitlichen Styling.</li>
+                    <li>ClosedText item type mit automatischer Korrektur.</li>
+                    <li>Statistiken werden in unterschiedlichen Bereichen im Studenten Dashboard erhoben.</li>
+                    <li>Statistiken werden automatisch ausgewertet und im Dashboard angezeigt.</li>
+                  </ul>
+                </p><hr class="my-3">
+                <p>
+                  Was soll noch umgesetzt werden:
+                  <ul class="list-disc list-inside">
+                    <li>Automatische Analysen bezüglich der verschiedenen Module.</li>
+                    <li>Vorschläge an den Prof basierend auf den Analysen.</li>
+                    <li>Weitere Statistiken erheben und evtl. Trends schon früh erkennen können.</li>
+                    <li>Dashboard weiter ausbauen.</li>
+                  </ul>
+                </p>
+              </div>
+            </div>
+          </div>
+          <div class="bg-white shadow sm:rounded-lg">
+            <div class="px-4 py-5 sm:p-6">
+              <h3 class="text-lg leading-6 font-medium text-gray-900">
+                Max
+              </h3>
+              <div class="mt-2 max-w-xl text-gray-600">
+                <p />
+              </div>
+            </div>
+          </div>
+          <div class="bg-white shadow sm:rounded-lg">
+            <div class="px-4 py-5 sm:p-6">
+              <h3 class="text-lg leading-6 font-medium text-gray-900">
+                Romy
+              </h3>
+              <div class="mt-2 max-w-xl text-gray-600">
+                <p />
+              </div>
+            </div>
+          </div>
+          <div class="bg-white shadow sm:rounded-lg">
+            <div class="px-4 py-5 sm:p-6">
+              <h3 class="text-lg leading-6 font-medium text-gray-900">
+                Rupert
+              </h3>
+              <div class="mt-2 max-w-xl text-gray-600">
+                <p />
+              </div>
+            </div>
+          </div>
+          <div class="bg-white shadow sm:rounded-lg">
+            <div class="px-4 py-5 sm:p-6">
+              <h3 class="text-lg leading-6 font-medium text-gray-900">
+                Elias
+              </h3>
+              <div class="mt-2 max-w-xl text-gray-600">
+                <p />
+              </div>
+            </div>
+          </div>
+        </div>
+      <!-- /End replace -->
+      </div>
+    </div>
+  </dashboardsProf-dashboard>
+</template>
+
+<script setup>
+</script>